Added new page ButtonsPage to have examples of buttons. Added new page to navigation stack.
