+++ ReactNativeKickOffCLI/screens/WelcomePage.tsx
@@ -1,6 +1,5 @@
 import { View, Button, Alert } from "react-native"
 import WelcomePoster from "../components/WelcomePoster"
-import CustomButton from "../components/CustomButton"
 import { NativeStackScreenProps } from "@react-navigation/native-stack";
 import { RootStackParamList } from "../App";
 import { SafeAreaView } from "react-native-safe-area-context";
@@ -15,34 +14,9 @@ const WelcomePage = ({navigation}: WelcomePageNavigationProp) => {
       <SafeAreaView>
         <View style={{padding: 16}}>
           <WelcomePoster name='Batman' tagline='Because, I am BATMAN'/>
-          <View style={{flexDirection: 'row'}}>
-            <Button 
-              onPress={showAlert} 
-              title='Button One'
-              color={'red'}
-            />
-            <Button 
-              onPress={showAnotherAlert} 
-              title='Button Two'
-              color={'green'}
-            />
-          </View>
-          <CustomButton title='Custom Button' onPress={showAlertForCustomButton}/>
         </View>
       </SafeAreaView>
     )
-}
-
-const showAlert = () => {
-    Alert.alert("Pressed button one")
-}
-
-function showAnotherAlert() {
-    Alert.alert("Pressed button two")
-}
-
-function showAlertForCustomButton() {
-    Alert.alert("Pressed custom button")
 }
 
 export default WelcomePage;
